parse more fields from the entries API
i needed the list of entries which is available only in the _embedded
field. while i'm here, just find all fields already. this will grow
the memory usage but should be negligible because we read the whole
output in memory already anyways.

i believe the proper optimization would be to stream entries through a
channel, but the JSON data structure doesn't lend itself to streaming
very well.